pyproject.toml
[project]
name = "hannou"
version = "0.1.0"
description = "Add your description here"
readme = "README"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.6",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
